+++ edc-extensions/agreements/retirement-evaluation-core/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -17,5 +17,5 @@
 #  SPDX-License-Identifier: Apache-2.0
 #################################################################################
 
-org.eclipse.tractusx.edc.agreements.retirement.AgreementsRetirementFunctionExtension
+org.eclipse.tractusx.edc.agreements.retirement.AgreementsRetirementPreValidatorRegisterExtension
 org.eclipse.tractusx.edc.agreements.retirement.defaults.DefaultAgreementRetirementProviderExtension
